Fixed #37142 -- Moved django_file_prefixes() to django.utils.warnings.
