+++ django/conf/__init__.py
@@ -16,10 +16,10 @@ from django.conf import global_settings
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.deprecation import (
     RemovedInDjango70Warning,
-    django_file_prefixes,
     warn_about_external_use,
 )
 from django.utils.functional import LazyObject, empty
+from django.utils.warnings import django_file_prefixes
 
 ENVIRONMENT_VARIABLE = "DJANGO_SETTINGS_MODULE"
 DEFAULT_STORAGE_ALIAS = "default"
