+++ django/db/models/query.py
@@ -37,8 +37,9 @@ from django.db.models.utils import (
     resolve_callables,
 )
 from django.utils import timezone
-from django.utils.deprecation import RemovedInDjango70Warning, django_file_prefixes
+from django.utils.deprecation import RemovedInDjango70Warning
 from django.utils.functional import cached_property
+from django.utils.warnings import django_file_prefixes
 
 # The maximum number of results to fetch in a get() query.
 MAX_GET_RESULTS = 21
