+++ docs/internals/contributing/writing-code/submitting-patches.txt
@@ -352,7 +352,8 @@ previous behavior, or standalone items that are unnecessary or unused when the
 deprecation ends. For example::
 
     import warnings
-    from django.utils.deprecation import RemovedInDjangoXXWarning, django_file_prefixes
+    from django.utils.deprecation import RemovedInDjangoXXWarning
+    from django.utils.warnings import django_file_prefixes
 
 
     # RemovedInDjangoXXWarning.
